backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-genai>=1.0.0
google-adk>=1.0.0
httpx>=0.27.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
sse-starlette>=2.0.0
jinja2>=3.1.0
aiofiles>=24.1.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
asyncpg>=0.30.0
alembic>=1.13.0
arq>=0.26.0
boto3>=1.37.0
structlog>=24.0.0
PyJWT>=2.8.0
matplotlib>=3.8.0
Pillow>=10.0.0
cryptography>=43.0.0
opentelemetry-api>=1.28.0
opentelemetry-sdk>=1.28.0
opentelemetry-exporter-otlp-proto-http>=1.28.0
opentelemetry-instrumentation-fastapi>=0.49b0
opentelemetry-instrumentation-httpx>=0.49b0
opentelemetry-instrumentation-sqlalchemy>=0.49b0
